# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/dates": {
            "get": {
                "tags": [
                    "Date"
                ],
                "summary": "DatesInfo",
                "operationId": "b198eddefa7705dcf2b7d9e229b3c102",
                "parameters": [
                    {
                        "name": "dates[]",
                        "in": "query",
                        "description": "Array of dates",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "format": "date"
                            }
                        },
                        "example": [
                            "2020-10-27",
                            "2023-03-22"
                        ]
                    },
                    {
                        "name": "semester",
                        "in": "query",
                        "description": "Semester ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "dates": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-12-20 17:49:36"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-12-20 17:49:36"
                                                    },
                                                    "number_of_lesson": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "time": {
                                                        "type": "string",
                                                        "format": "time",
                                                        "example": "13:00"
                                                    },
                                                    "subject": {
                                                        "type": "string",
                                                        "example": "Math"
                                                    },
                                                    "cabinet": {
                                                        "type": "string",
                                                        "example": "21"
                                                    },
                                                    "teacher": {
                                                        "type": "string",
                                                        "example": "User Userov Userovich"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}